+++ src/components/article/article.component.ts
@@ -1,6 +1,6 @@
 import { LitElement, customElement, property, html } from 'lit-element';
 import { connect } from 'pwa-helpers/connect-mixin';
-import store from '../../core/store';
+import store from '../../store';
 import { RootState } from 'typesafe-actions';
 import { ArticleTemplate } from './article.templates';
 import { deleteArticle, deleteComment, addComment } from './article.actions';
