--- src/components/root/navigation/navigation.epic.ts
@@ -1,15 +1,0 @@
-// import { RootAction, isActionOf } from 'typesafe-actions';
-// import { Epic } from 'redux-observable';
-// import { navigate } from './navigation.actions';
-// import { filter, map } from 'rxjs/operators';
-// import { router } from '../root.router';
-
-// // export const navigationEpic: Epic<RootAction, RootAction> = action$ =>
-// //   action$.pipe(
-// //     filter(isActionOf(navigate)),
-// //     map(action => {
-// //       router.resolve({ pathname: location.pathname });
-// //       // return navigate(action.payload);
-// //     }),
-// //     // tslint:disable-next-line: deprecation
-// //   );
