+++ src/core/navigation/navigation.reducer.ts
